python/requirements.txt
Flask==2.3.3
flask-cors==4.0.0

# Azure SDKs
azure-ai-textanalytics==5.3.0
azure-ai-translation-document==1.0.0b1
azure-core==1.29.5
azure-storage-blob==12.19.0
azure-cognitiveservices-speech==1.34.0

# Audio and ML
librosa==0.10.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
tensorflow==2.18.0
keras>=3.5.0

# scikit-learn==1.1.2

# OpenAI & dotenv
openai==1.23.6
python-dotenv==1.0.1

# Others
requests==2.31.0
IPython==8.22.2
openpyxl==3.0.10
whisper
transformers
torch
